Experimental shadow cutscene fix (#81)
* Untested: Change event code, potential fix #73

Adds an explicit call to Track 00 (silence) to the event code for this scene to avoid the as-yet-unexplained issue of it requesting the Kids Run Through The City Corner theme instead.
Untested. Un-assembled, even. I just prototyped this based on madsiur's advice.

* Fixed comments.

Changed the address in the comments to be correct for the event code jump, and better explained the replacement event code.

* Zero current volume when muting MSU-1

* Further fixes to other cutscenes. Some bugs.

 - Music still plays for about a frame (2 at most) at the beginning of
scenes as the music change happens on scene load and the event code
happens after that.
 - Scene 2 begins with a noticable (but not very long) delay, due to
having to add a lot of NOPs to the event code to pad out what I had to
overwrite.
 - Silence at inns after Shadow dreams. Considered a necessary
compromise for the moment, see issue for details as to why.